% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Semantic_functions.R, R/Final_functions.R
\name{sentiment_search}
\alias{sentiment_search}
\title{Title Sentiments matching word in the text}
\usage{
sentiment_search(doc_words, sentiment_patterns)

sentiment_search(doc_words, sentiment_patterns)
}
\arguments{
\item{doc_words}{document word list}

\item{sentiment_patterns}{vectors of positive or negative pattern words (vector) to look for in the the text we are analyzing. Vectors enable an unknown prefix or suffix.}
}
\value{
list of the number of matched pattern in the text an the exact words.

list of the number of matched pattern in the text an the exact words.
}
\description{
Title Sentiments matching word in the text

Title Sentiments matching word in the text
}
\examples{
sentiment_search(word_doc, Negative)
sentiment_search(word_doc, Negative)
}
