% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticFunctions.R
\name{TextPrep}
\alias{TextPrep}
\title{TextPrep}
\usage{
TextPrep(file)
}
\arguments{
\item{vector}{with words as elements}

\item{vector2}{}
}
\value{
vector with lowercase words extracted from txt file as elements

calculates how many words in the vector appear in the original txt file

how many words of each vector appear in the original txt file and
the ratio between these
}
\description{
TextPrep

comparison

summary
}
