% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Semantic_functions.R, R/Final_functions.R
\name{analyze_sentiment}
\alias{analyze_sentiment}
\title{Title}
\usage{
analyze_sentiment(x, positive, negative)

analyze_sentiment(x, positive, negative)
}
\arguments{
\item{x}{Path to the file}

\item{positive}{vectors of positive pattern words (vector) to look for in the the text we are analyzing. These Vectors enable an unknown prefix or suffix by use of (*) after the root word.}

\item{negative}{vectors of negative pattern words (vector) to look for in the the text we are analyzing. These Vectors enable an unknown prefix or suffix by use of (*) after the root word.}
}
\value{
synthetised results as list with the tendency of the sentiment from the reader perspective

synthetised results as list with the tendency of the sentiment from the reader perspective
}
\description{
Title

Title
}
\examples{
analyze_sentiment("~/Examen_1/Example_negative.txt", Positive, Negative)
analyze_sentiment("~/Examen_1/Example_negative.txt", Positive, Negative)
}
