% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semantic_functions.R
\name{analyze_sentiment}
\alias{analyze_sentiment}
\title{analyze_sentiment}
\usage{
analyze_sentiment(x, positive, negative)
}
\arguments{
\item{x}{Path to the text to be analysed.}

\item{positive}{vectors of positive pattern words (vector) to look for in the the text we are analyzing. Vectors enable an unknown prefix or suffix.}

\item{negative}{vectors of negative pattern words (vector) to look for in the the text we are analyzing. Vectors enable an unknown prefix or suffix.}
}
\value{
synthesized results as list with the predictive tendency of the sentiment from the reader perspective
}
\description{
analyze_sentiment
}
\examples{
analyze_sentiment("Example_negative.txt", Positive, Negative)
}
