% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Semantic_functions.R
\name{Read_text}
\alias{Read_text}
\title{Title List of words to be analyzed}
\usage{
Read_text(x)
}
\arguments{
\item{x}{Path to the file}
}
\value{
list of words from the text to be analysed.
}
\description{
Title List of words to be analyzed
}
\examples{
Read_text("~/Documents/Courses/Intermediate_R_Data_Science_and_Visualization_Techniques_beyong_base_R/Examen_1/Example_negative.txt")
}
