% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semantic_functions.R
\name{Read_text}
\alias{Read_text}
\title{Read_text: Text_to_list_words}
\usage{
Read_text(x)
}
\arguments{
\item{x}{Path to the text to be analysed.}
}
\value{
List of words from the text
}
\description{
Read_text: Text_to_list_words
}
\examples{
Read_text("exampleNegative.txt")
}
