% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semantic_functions.R
\name{sentiment_search}
\alias{sentiment_search}
\title{sentiment_search: Word_search_and_match}
\usage{
sentiment_search(doc_words, sentiment_patterns)
}
\arguments{
\item{doc_words}{Words list from the text}

\item{sentiment_patterns}{vectors of positive or negative pattern words (vector) to look for in the the text we are analyzing. Vectors enable an unknown prefix or suffix.}
}
\value{
list of the number of matched pattern in the text.
}
\description{
sentiment_search: Word_search_and_match
}
\examples{
sentiment_search(word_doc, Negative)
}
